+++ spykeviewer/api.py
@@ -9,6 +9,8 @@ class __ConfigOptions:
         self.load_selection_on_start = True
         # Default load mode (0 - Regular, 1 - Lazy, 2 - Cached lazy)
         self.load_mode = 0
+        # Default cascading mode (False - Regular, True - Lazy)
+        self.lazy_cascading = False
         # Use Enter key for code completion in console
         self.codecomplete_console_enter = True
         # Use Enter key for code completion in editor
@@ -39,26 +41,36 @@ window = None
 app = None
 
 
-def start_plugin(name):
+def start_plugin(name, current=None, selections=None):
     """ Start first plugin with given name and return result of start()
     method. Raises a SpykeException if not exactly one plugins with
     this name exist.
 
     :param str name: The name of the plugin. Should not include the
         directory.
+    :param current: A DataProvider to use as current selection. If
+        ``None``, the regular current selection from the GUI is used.
+    :param list selections: A list of DataProvider objects to use as
+        selections. If ``None``, the regular selections from the GUI
+        are used.
     """
-    return window.start_plugin(name)
+    return window.start_plugin(name, current, selections)
 
 
-def start_plugin_remote(name):
+def start_plugin_remote(name, current=None, selections=None):
     """ Start first plugin with given name using the remote script.
     Raises a SpykeException if not exactly one plugins with
     this name exist.
 
     :param str name: The name of the plugin. Should not include the
         directory.
+    :param current: A DataProvider to use as current selection. If
+        ``None``, the regular current selection from the GUI is used.
+    :param list selections: A list of DataProvider objects to use as
+        selections. If ``None``, the regular selections from the GUI
+        are used.
     """
-    window.start_plugin_remote(name)
+    window.start_plugin_remote(name, current, selections)
 
 
 def get_plugin(name):