+++ spykeviewer/ui/main_window.py
@@ -9,6 +9,7 @@ import copy
 import pickle
 import platform
 import subprocess
+import time
 
 from PyQt4.QtGui import (QMainWindow, QMessageBox,
                          QApplication, QFileDialog, QInputDialog,
@@ -16,7 +17,7 @@ from PyQt4.QtGui import (QMainWindow, QMessageBox,
                          QPalette, QDesktopServices, QFont, QAction,
                          QPixmap, QFileSystemModel, QHeaderView,
                          QActionGroup)
-from PyQt4.QtCore import (Qt, pyqtSignature, SIGNAL, QMimeData,
+from PyQt4.QtCore import (Qt, pyqtSignature, SIGNAL, QMimeData, QTimer,
                           QSettings, QCoreApplication, QUrl)
 
 from spyderlib.widgets.internalshell import InternalShell
@@ -29,6 +30,7 @@ from spykeutils.plugin.data_provider import DataProvider
 from spykeutils.plugin.analysis_plugin import AnalysisPlugin
 from spykeutils.progress_indicator import CancelException
 from spykeutils import SpykeException
+from spykeutils.plot.helper import ProgressIndicatorDialog
 
 from .. import api
 from main_ui import Ui_MainWindow
@@ -36,10 +38,10 @@ from settings import SettingsWindow
 from filter_dock import FilterDock
 from filter_dialog import FilterDialog
 from filter_group_dialog import FilterGroupDialog
-from progress_indicator_dialog import ProgressIndicatorDialog
 from plugin_editor_dock import PluginEditorDock
 import ipython_connection as ipy
 from plugin_model import PluginModel
+from remote_thread import RemoteThread
 
 
 logger = logging.getLogger('spykeviewer')
@@ -81,12 +83,23 @@ class MainWindow(QMainWindow, Ui_MainWindow):
         self.setupUi(self)
         self.dir = os.getcwd()
 
-        # Load mode menu
+        # Threads providing output from remotely started plugins
+        self.process_threads = {}
+        self.remote_process_counter = 0
+        QTimer.singleShot(1000, self.clean_finished_process_threads)
+
+        # Lazy load mode menu
         self.load_actions = QActionGroup(self)
         self.load_actions.setExclusive(True)
         self.actionFull_Load.setActionGroup(self.load_actions)
         self.actionLazy_Load.setActionGroup(self.load_actions)
         self.actionCached_Lazy_Load.setActionGroup(self.load_actions)
+
+        # Cascading mode menu
+        self.cascade_actions = QActionGroup(self)
+        self.cascade_actions.setExclusive(True)
+        self.actionFull.setActionGroup(self.cascade_actions)
+        self.actionLazy.setActionGroup(self.cascade_actions)
 
         # Python console
         self.console = None
@@ -225,6 +238,11 @@ class MainWindow(QMainWindow, Ui_MainWindow):
             self.actionCached_Lazy_Load.trigger()
         else:
             self.actionFull_Load.trigger()
+
+        if api.config.lazy_cascading:
+            self.actionLazy.trigger()
+        else:
+            self.actionFull.trigger()
 
         self.update_splash_screen('Loading plugins...')
         self.reload_plugins()
@@ -1134,9 +1152,13 @@ class MainWindow(QMainWindow, Ui_MainWindow):
                     return None
         return ana
 
-    def _run_plugin(self, plugin):
+    def _run_plugin(self, plugin, current=None, selections=None):
+        if current is None:
+            current = self.provider
+        if selections is None:
+            selections = self.selections
         try:
-            return plugin.start(self.provider, self.selections)
+            return plugin.start(current, selections)
         except SpykeException, err:
             QMessageBox.critical(self, 'Error executing plugin', str(err))
         except CancelException:
@@ -1216,13 +1238,9 @@ class MainWindow(QMainWindow, Ui_MainWindow):
         if not plugin:
             return
 
-        selections = self.serialize_selections()
-        config = pickle.dumps(plugin.get_parameters())
-        name = type(plugin).__name__
-        path = self.current_plugin_path()
-        self.send_plugin_info(name, path, selections, config)
+        self._execute_remote_plugin(plugin)
 
-    def send_plugin_info(self, name, path, selections, config):
+    def send_plugin_info(self, name, path, selections, config, io_files):
         """ Send information to start a plugin to the configured remote
         script.
 
@@ -1230,17 +1248,53 @@ class MainWindow(QMainWindow, Ui_MainWindow):
         :param str path: Path of the plugin file
         :param str selections: Serialized selections to use
         :param str config: Pickled plugin configuration
+        :param list io_files: List of paths to required IO plugins.
         """
         # Save files to circumvent length limit for command line
-        selection_path = os.path.join(self.selection_path, '.temp.sel')
+        selection_path = os.path.join(
+            self.selection_path, '.temp_%f_.sel' % time.time())
         with open(selection_path, 'w') as f:
             f.write(selections)
 
         params = ['python', self.remote_script,
                   name, path, selection_path, '-cf', '-sf',
                   '-c', config, '-dd', AnalysisPlugin.data_dir]
+        if io_files:
+            params.append('-io')
+            params.extend(io_files)
         params.extend(api.config.remote_script_parameters)
-        subprocess.Popen(params)
+        p = subprocess.Popen(
+            params, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+        std = RemoteThread(p, self.remote_process_counter, False, self)
+        err = RemoteThread(p, self.remote_process_counter, True, self)
+        self.connect(std, SIGNAL("output(int, QString)"), self.output_std)
+        self.connect(err, SIGNAL("output(int, QString)"), self.output_err)
+        self.connect(err, SIGNAL("execution_complete(int)"),
+                     self.remote_plugin_done)
+        std.start()
+        err.start()
+        self.process_threads[self.remote_process_counter] = (std, err)
+        print '[#%d started]' % self.remote_process_counter
+        self.remote_process_counter += 1
+
+    def output_std(self, id_, line):
+        print '[#%d]' % id_, line
+
+    def output_err(self, id_, line):
+        sys.stderr.write(line + '\n')
+
+    def remote_plugin_done(self, id_):
+        print '[#%d done]' % id_
+
+    def clean_finished_process_threads(self):
+        """ Periodically checks if threads for remote plugin output are
+        still running, removes them otherwise.
+        """
+        for k in self.process_threads.keys():
+            t = self.process_threads[k]
+            if not t[0].isRunning() and not t[1].isRunning():
+                del self.process_threads[k]
+        QTimer.singleShot(1000, self.clean_finished_process_threads)
 
     @pyqtSignature("")
     def on_actionEdit_Startup_Script_triggered(self):
@@ -1314,7 +1368,7 @@ class MainWindow(QMainWindow, Ui_MainWindow):
 
         return plugins[0]
 
-    def start_plugin(self, name):
+    def start_plugin(self, name, current=None, selections=None):
         """ Start first plugin with given name and return result of start()
         method. Raises a SpykeException if not exactly one plugins with
         this name exist.
@@ -1338,9 +1392,9 @@ class MainWindow(QMainWindow, Ui_MainWindow):
                     raise SpykeException(
                         'Multiple plugins named "%s" exist!' % name)
 
-        return self._run_plugin(plugins[0])
+        return self._run_plugin(plugins[0], current, selections)
 
-    def start_plugin_remote(self, name):
+    def start_plugin_remote(self, name, current=None, selections=None):
         """ Start first plugin with given name remotely. Does not return
         any value. Raises a SpykeException if not exactly one plugins with
         this name exist.
@@ -1351,11 +1405,7 @@ class MainWindow(QMainWindow, Ui_MainWindow):
         if len(plugins) > 1:
             raise SpykeException('Multiple plugins named "%s" exist!' % name)
 
-        selections = self.serialize_selections()
-        config = pickle.dumps(plugins[0].get_parameters())
-        name = type(plugins[0]).__name__
-        path = plugins[0].source_file
-        self.send_plugin_info(name, path, selections, config)
+        self._execute_remote_plugin(plugins[0], current, selections)
 
     def on_file_available(self, available):
         """ Callback when availability of a file for a plugin changes.
@@ -1371,6 +1421,10 @@ class MainWindow(QMainWindow, Ui_MainWindow):
                                   self.plugin_paths, self)
 
         if settings.exec_() == settings.Accepted:
+            try:
+                self.clean_temporary_selection_files(self.selection_path)
+            except:
+                pass  # Does not matter if e.g. old directory does not exist
             self.selection_path = settings.selection_path()
             self.filter_path = settings.filter_path()
             self.remote_script = settings.remote_script()
@@ -1431,6 +1485,14 @@ class MainWindow(QMainWindow, Ui_MainWindow):
             paths.append('/'.join(reversed(path)))
         return paths
 
+    def clean_temporary_selection_files(self, path):
+        """ Remove temporary .temp_..._.sel files from a directory.
+        These files are written when executing plugins remotely.
+        """
+        for f in os.listdir(path):
+            if f.startswith('.temp_') and f.endswith('_.sel'):
+                os.remove(os.path.join(path, f))
+
     def closeEvent(self, event):
         """ Saves filters, plugin configs and GUI state.
         """
@@ -1448,6 +1510,7 @@ class MainWindow(QMainWindow, Ui_MainWindow):
 
         self.save_selections_to_file(
             os.path.join(self.selection_path, '.current.sel'))
+        self.clean_temporary_selection_files(self.selection_path)
 
         # Ensure that filters folder exists
         if not os.path.exists(self.filter_path):