+++ CHANGELOG.rst
@@ -1,11 +1,18 @@
-* Optional transparent lazy loading for supported IOs.
+Version 0.4.0
+-------------
+
+* Optional transparent lazy loading and lazy cascading for supported IOs.
 * Splash screen while loading the application.
 * Open files dialog as an alternative to the "Files" dock.
+* Remotely started plugins can have a graphical progress bar.
+* Remotely started plugins now show text output and errors on internal
+  console.
 * Filters are automatically deactivated on loading a selection if they
   prevent parts of it to be shown.
 * A modified plugin is automatically saved before it is sent to a remote
   script.
-* Pyhton files can be dragged onto the editor to open them.
+* New features for correlogram and spike waveform plugins.
+* Python files can be dragged onto the editor to open them.
 * Annotation editor accessible through API.
 * Files can be loaded through API.